(graph) Refactor all graph nodes into a new library
