+++ core/CMakeLists.txt
@@ -12,7 +12,7 @@ set(io_srcs
 set(type_srcs
     src/UVMap.cpp
     src/ITK2VTK.cpp
-    src/CompositeTransform.cpp
+    src/Transforms.cpp
 )
 
 set(util_srcs
@@ -31,20 +31,20 @@ set(srcs
 
 configure_file(src/Version.cpp.in Version.cpp)
 
-add_library(${module}
+add_library(rt_${module}
     ${srcs}
     ${io_srcs}
     ${type_srcs}
     ${util_srcs}
     ${CMAKE_CURRENT_BINARY_DIR}/Version.cpp
 )
-add_library("${namespace}${module}" ALIAS "${module}")
-target_include_directories(${module}
+add_library("${namespace}${module}" ALIAS "rt_${module}")
+target_include_directories(rt_${module}
     PUBLIC
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
         $<INSTALL_INTERFACE:include>
 )
-target_link_libraries(${module}
+target_link_libraries(rt_${module}
     PUBLIC
         Boost::filesystem
         opencv_core
@@ -66,16 +66,16 @@ target_link_libraries(${module}
         TIFF::TIFF
 )
 if(TARGET ITKSmoothing)
-    target_link_libraries(${module} PRIVATE ITKSmoothing)
+    target_link_libraries(rt_${module} PRIVATE ITKSmoothing)
 endif()
-target_compile_features(${module} PUBLIC cxx_std_14)
-set_target_properties(${module} PROPERTIES
+target_compile_features(rt_${module} PUBLIC cxx_std_14)
+set_target_properties(rt_${module} PROPERTIES
     VERSION "${PROJECT_VERSION}"
     EXPORT_NAME "${module}"
 )
 
 install(
-    TARGETS "${module}"
+    TARGETS "rt_${module}"
     EXPORT "${targets_export_name}"
     ARCHIVE DESTINATION "lib"
     LIBRARY DESTINATION "lib"
