+++ core/include/rt/AffineLandmarkRegistration.hpp
@@ -2,13 +2,9 @@
 
 /** @file */
 
-#include <boost/filesystem.hpp>
-#include <smgl/Node.hpp>
-#include <smgl/Ports.hpp>
+#include <itkAffineTransform.h>
 
-#include "rt/ITKImageTypes.hpp"
 #include "rt/LandmarkRegistrationBase.hpp"
-#include "rt/types/CompositeTransform.hpp"
 
 namespace rt
 {
@@ -33,32 +29,4 @@ private:
     /** Transform */
     Transform::Pointer output_;
 };
-
-namespace graph
-{
-
-class AffineLandmarkRegistrationNode : public smgl::Node
-{
-    using Path = boost::filesystem::path;
-    using Metadata = smgl::Metadata;
-    using Transform = AffineLandmarkRegistration::Transform;
-
-public:
-    AffineLandmarkRegistrationNode();
-
-    smgl::InputPort<LandmarkContainer> fixedLandmarks{&fixed_};
-    smgl::InputPort<LandmarkContainer> movingLandmarks{&moving_};
-    smgl::OutputPort<CommonTransform::Pointer> transform{&tfm_};
-
-private:
-    AffineLandmarkRegistration reg_;
-    LandmarkContainer fixed_;
-    LandmarkContainer moving_;
-    CommonTransform::Pointer tfm_;
-
-    Metadata serialize_(bool useCache, const Path& cacheDir) override;
-    void deserialize_(const Metadata& meta, const Path& cacheDir) override;
-};
-
-}  // namespace graph
 }  // namespace rt
