+++ core/include/rt/DeformableRegistration.hpp
@@ -4,10 +4,6 @@
 
 #include <itkBSplineTransform.h>
 #include <opencv2/core.hpp>
-#include <smgl/Node.hpp>
-#include <smgl/Ports.hpp>
-
-#include "rt/types/CompositeTransform.hpp"
 
 namespace rt
 {
@@ -74,31 +70,4 @@ private:
     /** Stop condition if change in metric is less than this value */
     double gradientMagnitudeTolerance_{DEFAULT_GRAD_MAG_TOLERANCE};
 };
-
-namespace graph
-{
-
-class DeformableRegistrationNode : public smgl::Node
-{
-    using Transform = DeformableRegistration::Transform;
-
-public:
-    DeformableRegistrationNode();
-
-    smgl::InputPort<cv::Mat> fixedImage{
-        &reg_, &DeformableRegistration::setFixedImage};
-    smgl::InputPort<cv::Mat> movingImage{
-        &reg_, &DeformableRegistration::setMovingImage};
-    smgl::InputPort<int> iterations{
-        &reg_, &DeformableRegistration::setNumberOfIterations};
-    smgl::OutputPort<CommonTransform::Pointer> transform{&tfm_};
-
-private:
-    DeformableRegistration reg_;
-    CommonTransform::Pointer tfm_;
-
-    // TODO: Serialization
-};
-
-}  // namespace graph
 }  // namespace rt
