+++ core/include/rt/io/LandmarkReader.hpp
@@ -4,8 +4,6 @@
 
 #include <boost/filesystem.hpp>
 #include <opencv2/core.hpp>
-#include <smgl/Node.hpp>
-#include <smgl/Ports.hpp>
 
 #include "rt/LandmarkRegistrationBase.hpp"
 
@@ -49,29 +47,4 @@ private:
     LandmarkContainer moving_;
 };
 
-namespace graph
-{
-
-class LandmarkReaderNode : public smgl::Node
-{
-    using Path = boost::filesystem::path;
-    using Metadata = smgl::Metadata;
-
-public:
-    LandmarkReaderNode();
-    smgl::InputPort<Path> path{&path_};
-    smgl::OutputPort<LandmarkContainer> fixedLandmarks{&fixed_};
-    smgl::OutputPort<LandmarkContainer> movingLandmarks{&moving_};
-
-private:
-    LandmarkReader reader_;
-    Path path_;
-    LandmarkContainer fixed_;
-    LandmarkContainer moving_;
-
-    Metadata serialize_(bool /*unused*/, const Path& /*unused*/) override;
-    void deserialize_(const Metadata& meta, const Path& /*unused*/) override;
-};
-
-}  // namespace graph
 }  // namespace rt
