+++ core/include/rt/io/LandmarkWriter.hpp
@@ -3,8 +3,6 @@
 /** @file */
 
 #include <boost/filesystem.hpp>
-#include <smgl/Node.hpp>
-#include <smgl/Ports.hpp>
 
 #include "rt/LandmarkRegistrationBase.hpp"
 
@@ -54,29 +52,4 @@ private:
     LandmarkContainer moving_;
 };
 
-namespace graph
-{
-
-class LandmarkWriterNode : public smgl::Node
-{
-    using Path = boost::filesystem::path;
-    using Metadata = smgl::Metadata;
-
-public:
-    LandmarkWriterNode();
-    smgl::InputPort<Path> path{&path_};
-    smgl::InputPort<LandmarkContainer> fixed{&fixed_};
-    smgl::InputPort<LandmarkContainer> moving{&moving_};
-
-private:
-    LandmarkWriter writer_;
-    Path path_;
-    LandmarkContainer fixed_;
-    LandmarkContainer moving_;
-
-    Metadata serialize_(bool /*unused*/, const Path& /*unused*/) override;
-    void deserialize_(const Metadata& meta, const Path& /*unused*/) override;
-};
-
-}  // namespace graph
 }  // namespace rt
